@@ -8,6 +8,14 @@ using Valour.Shared;
 using Valour.Database.Items.Authorization;
 using Valour.Database.Items.Planets.Channels;
 using Valour.Shared.Items;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Http;
+using Valour.Shared.Http;
+using Valour.Database.Attributes;
+using System.Web.Mvc;
+using Valour.Database.Extensions;
+using Microsoft.Extensions.Logging;
 
 /*  Valour - A free and secure chat client
  *  Copyright (C) 2021 Vooper Media LLC
@@ -21,7 +29,7 @@ namespace Valour.Database.Items.Planets.Members;
 /// This class exists to add server funtionality to the PlanetMember
 /// class.
 /// </summary>
-public class PlanetMember : Item, IPlanetItemAPI<PlanetMember>, ISharedPlanetMember, INodeSpecific
+public class PlanetMember : PlanetItem, ISharedPlanetMember
 {
 
     public const int FLAG_UPDATE_ROLES = 0x01;
@@ -31,10 +39,6 @@ public class PlanetMember : Item, IPlanetItemAPI<PlanetMember>, ISharedPlanetMem
     [JsonIgnore]
     public virtual User User { get; set; }
 
-    [ForeignKey("Planet_Id")]
-    [JsonIgnore]
-    public virtual Planet Planet { get; set; }
-
     [InverseProperty("Member")]
     [JsonIgnore]
     public virtual ICollection<PlanetRoleMember> RoleMembership { get; set; }
@@ -43,11 +47,6 @@ public class PlanetMember : Item, IPlanetItemAPI<PlanetMember>, ISharedPlanetMem
     /// The user within the planet
     /// </summary>
     public ulong User_Id { get; set; }
-
-    /// <summary>
-    /// The planet the user is within
-    /// </summary>
-    public ulong Planet_Id { get; set; }
 
     /// <summary>
     /// The name to be used within the planet
@@ -70,7 +69,7 @@ public class PlanetMember : Item, IPlanetItemAPI<PlanetMember>, ISharedPlanetMem
     /// <summary>
     /// Returns all of the roles for a planet user
     /// </summary>
-    public async Task<List<PlanetRole>> GetRolesAsync(ValourDB db = null)
+    public async Task<List<PlanetRole>> GetRolesAsync(ValourDB db)
     {
         List<PlanetRole> roles;
 
@@ -87,31 +86,17 @@ public class PlanetMember : Item, IPlanetItemAPI<PlanetMember>, ISharedPlanetMem
     /// <summary>
     /// Loads role membership data from database
     /// </summary>
-    public async Task LoadRoleMembershipAsync(ValourDB db = null)
-    {
-        bool createdb = false;
-        if (db == null)
-        {
-            db = new ValourDB(ValourDB.DBOptions);
-            createdb = true;
-        }
-
+    public async Task LoadRoleMembershipAsync(ValourDB db) =>
         await db.Attach(this).Collection(x => x.RoleMembership)
                                  .Query()
                                  .Include(x => x.Role)
                                  .OrderBy(x => x.Role.Position)
                                  .LoadAsync();
 
-        if (createdb)
-        {
-            await db.DisposeAsync();
-        }
-    }
-
     /// <summary>
     /// Returns the member's primary role
     /// </summary>
-    public async Task<PlanetRole> GetPrimaryRoleAsync(ValourDB db = null)
+    public async Task<PlanetRole> GetPrimaryRoleAsync(ValourDB db)
     {
         if (RoleMembership == null)
         {
@@ -126,7 +111,7 @@ public class PlanetMember : Item, IPlanetItemAPI<PlanetMember>, ISharedPlanetMem
     /// </summary>
     public async Task<bool> HasPermissionAsync(PlanetPermission permission, ValourDB db)
     {
-        Planet ??= await db.Planets.FindAsync(Planet_Id);
+        await GetPlanetAsync(db);
         return await Planet.HasPermissionAsync(this, permission, db);
     }
 
@@ -161,7 +146,7 @@ public class PlanetMember : Item, IPlanetItemAPI<PlanetMember>, ISharedPlanetMem
                 var cperm = pair.perm as ChatChannelPermission;
                 var channel = pair.target as PlanetChatChannel;
 
-                if (!await channel.HasPermission(this, cperm, db))
+                if (!await channel.HasPermissionAsync(this, cperm, db))
                     return new TaskResult(false, "Member lacks " + cperm.Name + " channel permission.");
             }
             else if (pair.perm is CategoryPermission)
@@ -169,7 +154,7 @@ public class PlanetMember : Item, IPlanetItemAPI<PlanetMember>, ISharedPlanetMem
                 var cperm = pair.perm as CategoryPermission;
                 var channel = pair.target as PlanetCategoryChannel;
 
-                if (!await channel.HasPermission(this, cperm, db))
+                if (!await channel.HasPermissionAsync(this, cperm, db))
                     return new TaskResult(false, "Member lacks " + cperm.Name + " category permission.");
             }
             else
@@ -184,35 +169,13 @@ public class PlanetMember : Item, IPlanetItemAPI<PlanetMember>, ISharedPlanetMem
     /// <summary>
     /// Returns the user (async)
     /// </summary>
-    public async Task<User> GetUserAsync()
+    public async Task<User> GetUserAsync(ValourDB db) => 
+        User ??= await db.Users.FindAsync(User_Id);
+
+    public async Task<ulong> GetAuthorityAsync(ValourDB db)
     {
-        using (ValourDB db = new ValourDB(ValourDB.DBOptions))
-        {
-            return await db.Users.FindAsync(User_Id);
-        }
-    }
-
-    /// <summary>
-    /// Returns the planet (async)
-    /// </summary>
-    public async Task<Planet> GetPlanetAsync()
-    {
-        if (Planet != null) return Planet;
-
-        using (ValourDB db = new ValourDB(ValourDB.DBOptions))
-        {
-            Planet = await db.Planets.FindAsync(Planet_Id);
-        }
-
-        return Planet;
-    }
-
-    public async Task<ulong> GetAuthorityAsync()
-    {
-        if (Planet == null)
-        {
-            Planet = await GetPlanetAsync();
-        }
+        if (Planet is null)
+            await GetPlanetAsync(db);
 
         if (Planet.Owner_Id == User_Id)
         {
@@ -221,88 +184,309 @@ public class PlanetMember : Item, IPlanetItemAPI<PlanetMember>, ISharedPlanetMem
         }
         else
         {
-            var primaryRole = await GetPrimaryRoleAsync();
-
+            var primaryRole = await GetPrimaryRoleAsync(db);
             return primaryRole.GetAuthority();
         }
     }
 
-    public async Task<TaskResult> CanDeleteAsync(PlanetMember member, ValourDB db)
+    public async Task DeleteAsync(ValourDB db)
     {
-        // Needs to be able to GET in order to do anything else
-        var canGet = await ((IPlanetItemAPI<PlanetMember>)this).CanGetAsync(member, db);
-        if (!canGet.Success)
-            return canGet;
-
-        Planet ??= await GetPlanetAsync();
-
-        // Can always remove self
-        if (Id != member.Id)
-        {
-            // Need kick or ban to remove anyone else (ban implicitly grants kick)
-            if (!(await Planet.HasPermissionAsync(member, PlanetPermissions.Kick, db) ||
-                  await Planet.HasPermissionAsync(member, PlanetPermissions.Ban, db)))
-            {
-                return new TaskResult(false, "Member lacks planet permission " +
-                    PlanetPermissions.Kick.Name + " or " + PlanetPermissions.Ban.Name);
-            }
-        }
-
-        return new TaskResult(true, "Success");
-    }
-
-    public async Task<TaskResult> CanUpdateAsync(PlanetMember member, ValourDB db)
-    {
-        // Needs to be able to GET in order to do anything else
-        var canGet = await ((IPlanetItemAPI<PlanetMember>)this).CanGetAsync(member, db);
-        if (!canGet.Success)
-            return canGet;
-
-        // Only a member can edit themselves
-        if (Id != member.Id)
-        {
-            return new TaskResult(false, "Cannot edit another member.");
-        }
-
-        return new TaskResult(true, "Success");
-    }
-
-    public async Task<TaskResult> CanCreateAsync(PlanetMember member, ValourDB db)
-    {
-        // TODO: Be more specific
-        return await Task.FromResult(
-            new TaskResult(false, "Membership is created through the Planet API")
+        await db.BulkDeleteAsync(
+            db.PlanetRoleMembers.Where(x => x.Member_Id == Id)
         );
+
+        db.PlanetMembers.Remove(this);
     }
 
-    public async Task<TaskResult> ValidateItemAsync(PlanetMember old, ValourDB db)
+    // Helpful route to return the member for the authorizing user
+    [ValourRoute(HttpVerbs.Get, "/self"), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    public static void GetSelfRoute(HttpContext ctx) =>
+        Results.Json(ctx.GetMember());
+
+    [ValourRoute(HttpVerbs.Get), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    public static async Task<IResult> GetRoute(HttpContext ctx, ulong id)
     {
-        // This means this is an update rather
-        // than a creation
-        if (old is not null)
+        var db = ctx.GetDb();
+        var member = await FindAsync<PlanetMember>(id, db);
+
+        if (member is null)
+            return ValourResult.NotFound<PlanetMember>();
+
+        return Results.Json(member);
+    }
+
+    [ValourRoute(HttpVerbs.Get, "/byuser/{user_id}"), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    public static async Task<IResult> GetRoute(HttpContext ctx, ulong planet_id, ulong user_id)
+    {
+        var db = ctx.GetDb();
+        var member = await FindAsync(user_id, planet_id, db);
+
+        if (member is null)
+            return ValourResult.NotFound<PlanetMember>();
+
+        return Results.Json(member);
+    }
+
+    [ValourRoute(HttpVerbs.Post), TokenRequired, InjectDB]
+    public static async Task<IResult> PostRouteAsync(HttpContext ctx, ulong planet_id, string invite_code, [FromBody] PlanetMember member,
+        ILogger<PlanetMember> logger)
+    {
+        var db = ctx.GetDb();
+        var token = ctx.GetToken();
+
+        if (member.Planet_Id != planet_id)
+            return Results.BadRequest("Planet_Id does not match.");
+        if (member.User_Id != token.User_Id)
+            return Results.BadRequest("User_Id does not match.");
+
+        var nameValid = ValidateName(member);
+        if (!nameValid.Success)
+            return Results.BadRequest(nameValid.Message);
+
+        // Clear out pfp, it *must* be done through VMPS
+        member.Member_Pfp = null;
+
+        // Ensure member does not already exist
+        if (await db.PlanetMembers.AnyAsync(x => x.Planet_Id == planet_id && x.User_Id == token.User_Id))
+            return Results.BadRequest("Planet member already exists.");
+
+        var planet = await FindAsync<Planet>(planet_id, db);
+
+        if (!planet.Public)
         {
-            // Cannot change user ID
-            if (old.User_Id != User_Id)
-            {
-                return new TaskResult(false, "User_Id cannot be changed.");
-            }
+            if (invite_code is null)
+                return ValourResult.Forbid("The planet is not public. Please include invite_code.");
+
+            if (!await db.PlanetInvites.AnyAsync(x => x.Code == invite_code && x.Planet_Id == planet_id && DateTime.UtcNow > x.Created))
+                return ValourResult.Forbid("The invite code is invalid or expired.");
         }
 
+        try
+        {
+            await db.AddAsync(member);
+            await db.SaveChangesAsync();
+        }
+        catch (System.Exception e)
+        {
+            logger.LogError(e.Message);
+            return Results.Problem(e.Message);
+        }
+
+        PlanetHub.NotifyPlanetItemChange(member);
+
+        return Results.Created(member.GetUri(), member);
+    }
+
+    [ValourRoute(HttpVerbs.Put), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    public static async Task<IResult> PutRouteAsync(HttpContext ctx, ulong id, ulong planet_id, [FromBody] PlanetMember member,
+        ILogger<PlanetMember> logger)
+    {
+        var db = ctx.GetDb();
+        var token = ctx.GetToken();
+
+        var old = await FindAsync<PlanetMember>(id, db);
+
+        if (old is null)
+            return ValourResult.NotFound<PlanetMember>();
+
+        if (old.Id != member.Id)
+            return Results.BadRequest("Cannot change Id.");
+
+        if (token.User_Id != member.User_Id)
+            return Results.BadRequest("You can only modify your own membership.");
+
+        if (member.Planet_Id != planet_id)
+            return Results.BadRequest("Cannot change Planet_Id.");
+
+        if (old.Member_Pfp != member.Member_Pfp)
+            return Results.BadRequest("Cannot directly change pfp. Use VMPS.");
+
+        var nameValid = ValidateName(member);
+        if (!nameValid.Success)
+            return Results.BadRequest(nameValid.Message);
+
+        try
+        {
+            db.PlanetMembers.Update(member);
+            await db.SaveChangesAsync();
+        }
+        catch (System.Exception e)
+        {
+            logger.LogError(e.Message);
+            return Results.Problem(e.Message);
+        }
+
+        return Results.Ok(member);
+    }
+
+    [ValourRoute(HttpVerbs.Delete), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    public static async Task<IResult> DeleteRouteAsync(HttpContext ctx, ulong id,
+        ILogger<PlanetMember> logger)
+    {
+        var db = ctx.GetDb();
+        var authMember = ctx.GetMember();
+        var targetMember = await FindAsync<PlanetMember>(id, db);
+
+        if (authMember.Id != id)
+        {
+            if (!await authMember.HasPermissionAsync(PlanetPermissions.Kick, db))
+                return ValourResult.LacksPermission(PlanetPermissions.Kick);
+
+            if (await authMember.GetAuthorityAsync(db) < await targetMember.GetAuthorityAsync(db))
+                return ValourResult.Forbid("You have less authority than the target member.");
+        }
+            
+
+        var tran = await db.Database.BeginTransactionAsync();
+
+        try
+        {
+            await targetMember.DeleteAsync(db);
+            await db.SaveChangesAsync();
+        }
+        catch(System.Exception e)
+        {
+            logger.LogError(e.Message);
+            return Results.Problem(e.Message);
+        }
+
+        await tran.CommitAsync();
+        PlanetHub.NotifyPlanetItemDelete(targetMember);
+
+        return Results.NoContent();
+    }
+
+    public static TaskResult ValidateName(PlanetMember member)
+    {
         // Ensure nickname is valid
-        if (Nickname.Length > 32)
+        if (member.Nickname.Length > 32)
         {
             return new TaskResult(false, "Maximum nickname is 32 characters.");
         }
 
-        if (Member_Pfp != old.Member_Pfp)
+        return TaskResult.SuccessResult;
+    }
+
+    [ValourRoute(HttpVerbs.Get, "/{id}/roles"), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    public async Task<IResult> GetAllRolesForMember(HttpContext ctx, ulong id, ulong planet_id)
+    {
+        var db = ctx.GetDb();
+
+        var member = await db.PlanetMembers.Include(x => x.RoleMembership.OrderBy(x => x.Role.Position))
+                                           .ThenInclude(x => x.Role)
+                                           .FirstOrDefaultAsync(x => x.Id == id && x.Planet_Id == planet_id);
+        if (member is null)
+            return ValourResult.NotFound<PlanetMember>();
+
+        return Results.Json(member.RoleMembership.Select(r => r.Role_Id));
+    }
+
+    [ValourRoute(HttpVerbs.Post, "/{id}/roles/{role_id}"), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    public async Task<IResult> AddRoleToMember(HttpContext ctx, ulong id, ulong planet_id, ulong role_id,
+        ILogger<PlanetMember> logger)
+    {
+        var db = ctx.GetDb();
+        var authMember = ctx.GetMember();
+
+        var member = await FindAsync<PlanetMember>(id, db);
+        if (member is null)
+            return ValourResult.NotFound<PlanetMember>();
+
+        if (member.Planet_Id != planet_id)
+            return ValourResult.NotFound<PlanetMember>();
+
+        if (!await authMember.HasPermissionAsync(PlanetPermissions.ManageRoles, db))
+            return ValourResult.LacksPermission(PlanetPermissions.ManageRoles);
+
+        if (await db.PlanetRoleMembers.AnyAsync(x => x.Member_Id == member.Id && x.Role_Id == role_id))
+            return Results.BadRequest("The member already has this role");
+
+        var role = await db.PlanetRoles.FindAsync(role_id);
+        if (role is null)
+            return ValourResult.NotFound<PlanetRole>();
+
+        var authAuthority = await authMember.GetAuthorityAsync(db);
+        if (role.GetAuthority() > authAuthority)
+            return ValourResult.Forbid("You have lower authority than the role you are trying to add");
+
+        PlanetRoleMember newRoleMember = new()
         {
-            // TODO: Automatically use VMPS
-            return new TaskResult(false, "Profile picture must be changed through Upload API.");
+            Id = IdManager.Generate(),
+            Member_Id = member.Id,
+            Role_Id = role_id,
+            User_Id = member.User_Id
+        };
+
+        try
+        {
+            await db.PlanetRoleMembers.AddAsync(newRoleMember);
+            await db.SaveChangesAsync();
+        }
+        catch (System.Exception e)
+        {
+            logger.LogError(e.Message);
+            return Results.Problem(e.Message);
         }
 
-        return await Task.FromResult(
-            new TaskResult(true, "Success")
-        );
+        PlanetHub.NotifyPlanetItemChange(newRoleMember);
+
+        return Results.Created(newRoleMember.GetUri(), newRoleMember);
+    }
+
+
+    [ValourRoute(HttpVerbs.Delete, "/{id}/roles/{role_id}"), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    public async Task<IResult> RemoveRoleFromMember(HttpContext ctx, ulong id, ulong planet_id, ulong role_id, [FromHeader] string authorization,
+        ILogger<PlanetMember> logger)
+    {
+
+        var db = ctx.GetDb();
+        var authMember = ctx.GetMember();
+
+        var member = await FindAsync<PlanetMember>(id, db);
+        if (member is null)
+            return ValourResult.NotFound<PlanetMember>();
+
+        if (member.Planet_Id != planet_id)
+            return ValourResult.NotFound<PlanetMember>();
+
+        if (!await authMember.HasPermissionAsync(PlanetPermissions.ManageRoles, db))
+            return ValourResult.LacksPermission(PlanetPermissions.ManageRoles);
+
+        var oldRoleMember = await db.PlanetRoleMembers.FirstOrDefaultAsync(x => x.Member_Id == member.Id && x.Role_Id == role_id);
+
+        if (oldRoleMember is null)
+            return Results.BadRequest("The member does not have this role");
+
+        var role = await db.PlanetRoles.FindAsync(role_id);
+        if (role is null)
+            return ValourResult.NotFound<PlanetRole>();
+
+        var authAuthority = await authMember.GetAuthorityAsync(db);
+        if (role.GetAuthority() > authAuthority)
+            return ValourResult.Forbid("You have less authority than the role you are trying to remove"); ;
+
+        try
+        {
+            db.PlanetRoleMembers.Remove(oldRoleMember);
+            await db.SaveChangesAsync();
+        }
+        catch (System.Exception e)
+        {
+            logger.LogError(e.Message);
+            return Results.Problem(e.Message);
+        }
+
+        PlanetHub.NotifyPlanetItemDelete(oldRoleMember);
+
+        return Results.NoContent();
     }
 }
 
