@@ -1,19 +1,29 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Text.Json.Serialization;
 using System.Text.RegularExpressions;
+using System.Web.Mvc;
+using Valour.Database.Attributes;
 using Valour.Database.Items.Authorization;
 using Valour.Database.Items.Planets.Members;
 using Valour.Shared;
 using Valour.Shared.Authorization;
+using Valour.Shared.Http;
 using Valour.Shared.Items;
 using Valour.Shared.Items.Authorization;
 using Valour.Shared.Items.Planets.Channels;
+using Valour.Database.Extensions;
+using Z.BulkOperations;
+using Valour.Database.Nodes;
+using Microsoft.Extensions.Logging;
 
 namespace Valour.Database.Items.Planets.Channels;
 
 [Table("PlanetCategoryChannels")]
-public class PlanetCategoryChannel : PlanetChannel, ISharedPlanetCategoryChannel, INodeSpecific
+public class PlanetCategoryChannel : PlanetChannel, ISharedPlanetCategoryChannel
 {
 
     [JsonIgnore]
@@ -26,91 +36,9 @@ public class PlanetCategoryChannel : PlanetChannel, ISharedPlanetCategoryChannel
     public override ItemType ItemType => ItemType.PlanetCategoryChannel;
 
     /// <summary>
-    /// Tries to delete the category while respecting constraints
-    /// </summary>
-    public override async Task DeleteAsync(ValourDB db)
-    {
-        // Remove permission nodes
-        db.PermissionsNodes.RemoveRange(
-            db.PermissionsNodes.Where(x => x.Target_Id == Id)
-        );
-
-        // Remove category
-        db.PlanetCategoryChannels.Remove(
-            await db.PlanetCategoryChannels.FindAsync(Id)
-        );
-
-        // Save changes
-        await db.SaveChangesAsync();
-
-        // Notify of update
-        PlanetHub.NotifyPlanetItemDelete(this);
-    }
-
-    /// <summary>
-    /// Validates the parent and position of this category
-    /// </summary>
-    public async Task<TaskResult> ValidateParentAndPosition(ValourDB db)
-    {
-        if (Parent_Id != null)
-        {
-            var parent = await db.PlanetCategoryChannels.FindAsync(Parent_Id);
-            if (parent == null) return new TaskResult(false, "Could not find parent");
-            if (parent.Planet_Id != Planet_Id) return new TaskResult(false, "Parent category belongs to a different planet");
-            if (parent.Id == Id) return new TaskResult(false, "Cannot be own parent");
-
-            // Automatically determine position in this case
-            if (Position < 0)
-            {
-                this.Position = (ushort)(await db.PlanetChannels.CountAsync(x => x.Parent_Id == Parent_Id));
-            }
-            else
-            {
-                // Ensure position is not already taken
-                if (!await HasUniquePosition(db))
-                    return new TaskResult(false, "The position is already taken.");
-            }
-
-            // Ensure this category does not contain itself
-            var loop_parent = parent;
-
-            while (loop_parent.Parent_Id != null)
-            {
-                if (loop_parent.Parent_Id == Id)
-                {
-                    return new TaskResult(false, "Cannot create parent loop.");
-                }
-
-                loop_parent = await db.PlanetCategoryChannels.FindAsync(loop_parent.Parent_Id);
-            }
-        }
-        else
-        {
-            if (Position < 0)
-            {
-                this.Position = (ushort)(await db.PlanetChannels.CountAsync(x => x.Planet_Id == Planet_Id && x.Parent_Id == null));
-            }
-            else
-            {
-                // Ensure position is not already taken
-                if (!await HasUniquePosition(db))
-                    return new TaskResult(false, "The position is already taken.");
-            }
-        }
-
-        return TaskResult.SuccessResult;
-    }
-
-    public async Task<bool> HasUniquePosition(ValourDB db) =>
-        // Ensure position is not already taken
-        !(await db.PlanetChannels.AnyAsync(x => x.Parent_Id == Parent_Id && // Same parent
-                                                x.Position == Position && // Same position
-                                                x.Id != Id)); // Not self
-
-    /// <summary>
     /// Returns if the member has the given permission in this category
     /// </summary>
-    public async Task<bool> HasPermission(PlanetMember member, Permission permission, ValourDB db)
+    public override async Task<bool> HasPermissionAsync(PlanetMember member, Permission permission, ValourDB db)
     {
         Planet planet = await GetPlanetAsync(db);
 
@@ -122,7 +50,7 @@ public class PlanetCategoryChannel : PlanetChannel, ISharedPlanetCategoryChannel
         // If true, we ask the parent
         if (InheritsPerms)
         {
-            return await (await GetParentAsync(db)).HasPermission(member, permission, db);
+            return await (await GetParentAsync(db)).HasPermissionAsync(member, permission, db);
         }
 
         var roles = await member.GetRolesAsync(db);
@@ -177,8 +105,284 @@ public class PlanetCategoryChannel : PlanetChannel, ISharedPlanetCategoryChannel
         return false;
     }
 
+    public async Task DeleteAsync(ValourDB db)
+    {
+        // Remove permission nodes
+        await db.BulkDeleteAsync(
+            db.PermissionsNodes.Where(x => x.Target_Id == Id)
+        );
+
+        // Remove category
+        db.PlanetCategoryChannels.Remove(
+            this
+        );
+    }
+
     /// <summary>
-    /// Validates that a given name is allowable for a server
+    /// Returns the children for this category
+    /// </summary>
+    public async Task<List<PlanetChannel>> GetChildrenAsync(ValourDB db)
+        => await db.PlanetChannels.Where(x => x.Parent_Id == Id).ToListAsync();
+
+    #region Routes
+
+    [ValourRoute(HttpVerbs.Get), TokenRequired, InjectDB]
+    [PlanetMembershipRequired, CategoryChannelPermsRequired(CategoryPermissionsEnum.View)]
+    public static IResult GetRoute(HttpContext ctx, ulong id) =>
+        Results.Json(ctx.GetItem<PlanetCategoryChannel>(id));
+
+    [ValourRoute(HttpVerbs.Put), TokenRequired, InjectDB]
+    [PlanetMembershipRequired, PlanetPermsRequired(PlanetPermissionsEnum.ManageCategories)]
+    [CategoryChannelPermsRequired(CategoryPermissionsEnum.View, 
+                                  CategoryPermissionsEnum.ManageCategory)]
+    public static async Task<IResult> PutRouteAsync(HttpContext ctx, ulong id, [FromBody] PlanetCategoryChannel category, 
+        ILogger<PlanetCategoryChannel> logger)
+    {
+        // Get resources
+        var db = ctx.GetDb();
+        var old = ctx.GetItem<PlanetCategoryChannel>(id);
+
+        // Validation
+        if (old.Id != category.Id)
+            return Results.BadRequest("Cannot change Id.");
+        if (old.Planet_Id != category.Planet_Id)
+            return Results.BadRequest("Cannot change Planet_Id.");
+
+        var nameValid = ValidateName(category.Name);
+        if (!nameValid.Success)
+            return Results.BadRequest(nameValid.Message);
+
+        var descValid = ValidateDescription(category.Description);
+        if (!descValid.Success)
+            return Results.BadRequest(descValid.Message);
+
+        var positionValid = await ValidateParentAndPosition(db, category);
+        if (!positionValid.Success)
+            return Results.BadRequest(positionValid.Message);
+
+        // Update
+        try
+        {
+            db.PlanetCategoryChannels.Update(category);
+            await db.SaveChangesAsync();
+        }
+        catch (System.Exception e)
+        {
+            logger.LogError(e.Message);
+            return Results.Problem(e.Message);
+        }
+
+        PlanetHub.NotifyPlanetItemChange(category);
+
+        // Response
+        return Results.Ok(category);
+    }
+
+    [ValourRoute(HttpVerbs.Post), TokenRequired, InjectDB]
+    [PlanetMembershipRequired, PlanetPermsRequired(PlanetPermissionsEnum.ManageCategories)]
+    public static async Task<IResult> PostRouteAsync(HttpContext ctx, ulong planet_id, [FromBody] PlanetCategoryChannel category,
+        ILogger<PlanetCategoryChannel> logger)
+    {
+        // Get resources
+        var db = ctx.GetDb();
+        var member = ctx.GetMember();
+
+        if (category.Planet_Id != planet_id)
+            return Results.BadRequest("Planet_Id mismatch.");
+
+        var nameValid = ValidateName(category.Name);
+        if (!nameValid.Success)
+            return Results.BadRequest(nameValid.Message);
+
+        var descValid = ValidateDescription(category.Description);
+        if (!descValid.Success)
+            return Results.BadRequest(descValid.Message);
+
+        var positionValid = await ValidateParentAndPosition(db, category);
+        if (!positionValid.Success)
+            return Results.BadRequest(positionValid.Message);
+
+        // Ensure user has permission for parent category management
+        if (category.Parent_Id is not null)
+        {
+            var parent_cat = await db.PlanetCategoryChannels.FindAsync(category.Parent_Id);
+            if (!await parent_cat.HasPermissionAsync(member, CategoryPermissions.ManageCategory, db))
+                return ValourResult.LacksPermission(CategoryPermissions.ManageCategory);
+        }
+
+        try
+        {
+            await db.PlanetCategoryChannels.AddAsync(category);
+            await db.SaveChangesAsync();
+        }
+        catch (System.Exception e)
+        {
+            logger.LogError(e.Message);
+            return Results.Problem(e.Message);
+        }
+
+        PlanetHub.NotifyPlanetItemChange(category);
+
+        return Results.Created(category.GetUri(), category);
+    }
+
+
+    [ValourRoute(HttpVerbs.Delete), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    [PlanetPermsRequired(PlanetPermissionsEnum.ManageCategories),
+     CategoryChannelPermsRequired(CategoryPermissionsEnum.View,
+                                  CategoryPermissionsEnum.ManageCategory)]
+    public static async Task<IResult> DeleteRouteAsync(HttpContext ctx, ulong id, ulong planet_id,
+        ILogger<PlanetCategoryChannel> logger)
+    {
+        var db = ctx.GetDb();
+        var category = ctx.GetItem<PlanetCategoryChannel>(id);
+
+        if (await db.PlanetCategoryChannels.CountAsync(x => x.Planet_Id == planet_id) < 2)
+            return Results.BadRequest("Last category cannot be deleted.");
+
+        var childCount = await db.PlanetChannels.CountAsync(x => x.Parent_Id == id);
+
+        if (childCount > 0)
+            return Results.BadRequest("Category must be empty.");
+
+        // Always use transaction for multi-step DB operations
+        using var transaction = await db.Database.BeginTransactionAsync();
+
+        try
+        {
+            await category.DeleteAsync(db);
+            await db.SaveChangesAsync();
+            await transaction.CommitAsync();
+        }
+        catch (System.Exception e)
+        {
+            logger.LogError(e.Message);
+            await transaction.RollbackAsync();
+            return Results.Problem(e.Message);
+        }
+
+        PlanetHub.NotifyPlanetItemDelete(category);
+
+        return Results.NoContent();
+    }
+
+    [ValourRoute(HttpVerbs.Get, "/children"), TokenRequired, InjectDB]
+    [PlanetMembershipRequired, CategoryChannelPermsRequired(CategoryPermissionsEnum.View)]
+    public async Task<IResult> GetChildrenRouteAsync(HttpContext ctx, ulong id)
+    {
+        var category = ctx.GetItem<PlanetCategoryChannel>(id);
+        var db = ctx.GetDb();
+
+        // Build child list. We don't have to check permissions for each, because even if the ID is there,
+        // it's impossible to get any details on the channels that are hidden.
+        var children_ids = await db.PlanetChannels.Where(x => x.Parent_Id == id).Select(x => x.Id).ToListAsync();
+
+        return Results.Json(children_ids);
+    }
+
+    [ValourRoute(HttpVerbs.Post, "/children/order"), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    [PlanetPermsRequired(PlanetPermissionsEnum.ManageCategories),
+     CategoryChannelPermsRequired(CategoryPermissionsEnum.ManageCategory)]
+    public static async Task<IResult> SetChildOrderRouteAsync(HttpContext ctx, ulong id, ulong planet_id, [FromBody] ulong[] order,
+        ILogger<PlanetCategoryChannel> logger)
+    {
+        var db = ctx.GetDb();
+        var category = ctx.GetItem<PlanetCategoryChannel>(id);
+
+        if (category.Planet_Id != planet_id)
+            return Results.BadRequest("Parent_Id mismatch.");
+
+        order = order.Distinct().ToArray();
+
+        var totalChildren = await db.PlanetChannels.CountAsync(x => x.Parent_Id == id);
+
+        if (totalChildren != order.Length)
+            return Results.BadRequest("Your order does not contain all the children.");
+
+        // Use transaction so we can stop at any failure
+        var tran = await db.Database.BeginTransactionAsync();
+
+        List<PlanetChannel> children = new();
+
+        try
+        {
+            var pos = 0;
+            foreach (var child_id in order)
+            {
+                var child = await FindAsync<PlanetChannel>(child_id, db);
+                if (child is null)
+                {
+                    return ValourResult.NotFound<PlanetChannel>();
+                }
+
+                if (child.Parent_Id != category.Id)
+                    return Results.BadRequest($"Category {child_id} is not a child of {category.Id}.");
+
+                child.Position = pos;
+
+                db.PlanetChannels.Update(child);
+
+                children.Add(child);
+
+                pos++;
+            }
+
+            await db.SaveChangesAsync();
+        }
+        catch (System.Exception e)
+        {
+            await tran.RollbackAsync();
+            logger.LogError(e.Message);
+            return Results.Problem(e.Message);
+        }
+
+        await tran.CommitAsync();
+
+        foreach (var child in children)
+        {
+            PlanetHub.NotifyPlanetItemChange(child);
+        }
+
+        return Results.NoContent();
+
+    }
+
+    /* This route isn't actually needed because you can PUT a new parent_id onto a category.
+     * But it's still a good reference.
+    
+    [ValourRoute(HttpVerbs.Post, "/children/{child_id}"), TokenRequired, InjectDB]
+    [PlanetMembershipRequired, PlanetPermsRequired(PlanetPermissionsEnum.ManageCategories),
+     CategoryChannelPermsRequired("id", CategoryPermissionsEnum.ManageCategory), // Need permission for parent
+     CategoryChannelPermsRequired("child_id", CategoryPermissionsEnum.ManageCategory)] // Need permission for child
+        public static async Task<IResult> AddChildRouteAsync(HttpContext ctx, ulong id, ulong planet_id, ulong child_id)
+        {
+            var db = ctx.GetDB();
+
+            var parent = ctx.GetItem<PlanetCategoryChannel>(id);
+            var child = ctx.GetItem<PlanetCategoryChannel>(child_id);
+
+            if (parent.Planet_Id != planet_id || child.Planet_Id != planet_id)
+                return Results.BadRequest("Planet_Id mismatch.");
+
+            child.Parent_Id = parent.Id;
+
+            db.PlanetChannels.Update(child);
+            await db.SaveChangesAsync();
+
+            PlanetHub.NotifyPlanetItemChange(child);
+
+            return Results.Ok(child);
+        }
+        */
+
+    #endregion
+
+    #region Validation
+
+    /// <summary>
+    /// Validates that a given name is allowable
     /// </summary>
     public static TaskResult ValidateName(string name)
     {
@@ -195,70 +399,73 @@ public class PlanetCategoryChannel : PlanetChannel, ISharedPlanetCategoryChannel
         return TaskResult.SuccessResult;
     }
 
-    public override async Task<TaskResult> CanGetAsync(PlanetMember member, ValourDB db)
+    /// <summary>
+    /// Validates that a given description is allowable
+    /// </summary>
+    public static TaskResult ValidateDescription(string desc)
     {
-        if (!await HasPermission(member, CategoryPermissions.View, db))
-            return new TaskResult(false, "Member lacks category permission " + CategoryPermissions.View.Name);
+        if (desc.Length > 500)
+        {
+            return new TaskResult(false, "Planet descriptions must be 500 characters or less.");
+        }
 
         return TaskResult.SuccessResult;
     }
 
-    public override async Task<TaskResult> CanUpdateAsync(PlanetMember member, PlanetItem old, ValourDB db)
+    /// <summary>
+    /// Validates the parent and position of this category
+    /// </summary>
+    public static async Task<TaskResult> ValidateParentAndPosition(ValourDB db, PlanetCategoryChannel category)
     {
-        var canCreate = await CanCreateAsync(member, db);
-        if (!canCreate.Success)
-            return canCreate;
+        if (category.Parent_Id != null)
+        {
+            var parent = await db.PlanetCategoryChannels.FindAsync(category.Parent_Id);
+            if (parent == null) return new TaskResult(false, "Could not find parent");
+            if (parent.Planet_Id != category.Planet_Id) return new TaskResult(false, "Parent category belongs to a different planet");
+            if (parent.Id == category.Id) return new TaskResult(false, "Cannot be own parent");
 
-        // Update needs specific perm for this category, everything else is same as create
-        if (!await HasPermission(member, CategoryPermissions.ManageCategory, db))
-            return new TaskResult(false, "Member lacks category permission " + CategoryPermissions.ManageCategory.Name);
+            // Automatically determine position in this case
+            if (category.Position < 0)
+            {
+                category.Position = (ushort)(await db.PlanetChannels.CountAsync(x => x.Parent_Id == category.Parent_Id));
+            }
+            else
+            {
+                // Ensure position is not already taken
+                if (!await HasUniquePosition(db, category))
+                    return new TaskResult(false, "The position is already taken.");
+            }
+
+            // Ensure this category does not contain itself
+            var loop_parent = parent;
+
+            while (loop_parent.Parent_Id != null)
+            {
+                if (loop_parent.Parent_Id == category.Id)
+                {
+                    return new TaskResult(false, "Cannot create parent loop.");
+                }
+
+                loop_parent = await db.PlanetCategoryChannels.FindAsync(loop_parent.Parent_Id);
+            }
+        }
+        else
+        {
+            if (category.Position < 0)
+            {
+                category.Position = (ushort)(await db.PlanetChannels.CountAsync(x => x.Planet_Id == category.Planet_Id && x.Parent_Id == null));
+            }
+            else
+            {
+                // Ensure position is not already taken
+                if (!await HasUniquePosition(db, category))
+                    return new TaskResult(false, "The position is already taken.");
+            }
+        }
 
         return TaskResult.SuccessResult;
     }
 
-    public override async Task<TaskResult> CanCreateAsync(PlanetMember member, ValourDB db)
-    {
-        var canGet = await CanGetAsync(member, db);
-        if (!canGet.Success)
-            return canGet;
-
-        await GetPlanetAsync(db);
-
-        if (!await Planet.HasPermissionAsync(member, PlanetPermissions.ManageCategories, db))
-            return new TaskResult(false, "Member lacks Planet Permission " + PlanetPermissions.ManageCategories.Name);
-
-        var nameValid = ValidateName(Name);
-        if (!nameValid.Success)
-            return nameValid;
-
-        var parentPosValid = await ValidateParentAndPosition(db);
-        if (!parentPosValid.Success)
-            return parentPosValid;
-
-        return TaskResult.SuccessResult;
-    }
-
-    public override async Task<TaskResult> CanDeleteAsync(PlanetMember member, ValourDB db)
-    {
-        await GetPlanetAsync(db);
-
-        if (!await Planet.HasPermissionAsync(member, PlanetPermissions.ManageCategories, db))
-            return new TaskResult(false, "Member lacks planet permission " + PlanetPermissions.ManageCategories.Name);
-
-        if (!await HasPermission(member, CategoryPermissions.ManageCategory, db))
-            return new TaskResult(false, "Member lacks category permission " + CategoryPermissions.ManageCategory.Name);
-
-
-        if (await db.PlanetCategoryChannels.CountAsync(x => x.Planet_Id == Planet_Id) < 2)
-            return new TaskResult(false, "Last category cannot be deleted");
-
-        var childCategoryCount = await db.PlanetCategoryChannels.CountAsync(x => x.Parent_Id == Id);
-        var childChannelCount = await db.PlanetChatChannels.CountAsync(x => x.Parent_Id == Id);
-
-        if (childCategoryCount != 0 || childChannelCount != 0)
-            return new TaskResult(false, "Category must be empty");
-
-        return new TaskResult(true, "Success");
-    }
+    #endregion
 }
 
