@@ -10,6 +10,15 @@ using Valour.Database.Items.Planets.Members;
 using Valour.Shared.Authorization;
 using Valour.Shared.Items.Planets.Channels;
 using Valour.Shared.Items;
+using Valour.Database.Attributes;
+using System.Web.Mvc;
+using Valour.Database.Extensions;
+using Microsoft.AspNetCore.Mvc;
+using Valour.Shared.Http;
+using Microsoft.Extensions.Logging;
+using Valour.Database.Items.Messages;
+using Valour.Database.Workers;
+using Valour.Shared.MPS;
 
 /*  Valour - A free and secure chat client
  *  Copyright (C) 2021 Vooper Media LLC
@@ -20,7 +29,7 @@ using Valour.Shared.Items;
 namespace Valour.Database.Items.Planets.Channels;
 
 [Table("PlanetChatChannels")]
-public class PlanetChatChannel : PlanetChannel, ISharedPlanetChatChannel, INodeSpecific
+public class PlanetChatChannel : PlanetChannel, ISharedPlanetChatChannel
 {
     public ulong MessageCount { get; set; }
 
@@ -38,7 +47,7 @@ public class PlanetChatChannel : PlanetChannel, ISharedPlanetChatChannel, INodeS
     /// <summary>
     /// Returns if a given member has a channel permission
     /// </summary>
-    public async Task<bool> HasPermission(PlanetMember member, Permission permission, ValourDB db)
+    public override async Task<bool> HasPermissionAsync(PlanetMember member, Permission permission, ValourDB db)
     {
         await GetPlanetAsync(db);
 
@@ -48,7 +57,7 @@ public class PlanetChatChannel : PlanetChannel, ISharedPlanetChatChannel, INodeS
         // If true, we just ask the category
         if (InheritsPerms)
         {
-            return await (await GetParentAsync(db)).HasPermission(member, permission, db);
+            return await (await GetParentAsync(db)).HasPermissionAsync(member, permission, db);
         }
 
 
@@ -102,8 +111,315 @@ public class PlanetChatChannel : PlanetChannel, ISharedPlanetChatChannel, INodeS
         return false;
     }
 
+    public async Task DeleteAsync(ValourDB db)
+    {
+        // Remove permission nodes
+        await db.PermissionsNodes.BulkDeleteAsync(
+            db.PermissionsNodes.Where(x => x.Target_Id == Id)
+        );
+
+        // Remove messages
+        await db.PlanetMessages.BulkDeleteAsync(
+            db.PlanetMessages.Where(x => x.Channel_Id == Id)
+        );
+
+        // Remove channel
+        db.PlanetChatChannels.Remove(this);
+    }
+
     /// <summary>
-    /// Validates that a given name is allowable for a channel
+    /// Returns all members who can see this channel
+    /// </summary>
+    public async Task<List<PlanetMember>> GetChannelMembersAsync(ValourDB db)
+    {
+        List<PlanetMember> members = new List<PlanetMember>();
+
+        var planetMembers = db.PlanetMembers.Include(x => x.RoleMembership).Where(x => x.Planet_Id == Planet_Id);
+
+        foreach (var member in planetMembers)
+        {
+            if (await HasPermissionAsync(member, ChatChannelPermissions.View, db))
+            {
+                members.Add(member);
+            }
+        }
+
+        return members;
+    }
+
+    #region Routes
+
+    [ValourRoute(HttpVerbs.Get), TokenRequired, InjectDB]
+    [PlanetMembershipRequired, ChatChannelPermsRequired(ChatChannelPermissionsEnum.View)]
+    public static IResult GetRoute(HttpContext ctx, ulong id) =>
+        Results.Json(ctx.GetItem<PlanetChatChannel>(id));
+
+    [ValourRoute(HttpVerbs.Post), TokenRequired, InjectDB]
+    [PlanetMembershipRequired, PlanetPermsRequired(PlanetPermissionsEnum.ManageChannels)]
+    public static async Task<IResult> PostRouteAsync(HttpContext ctx, ulong planet_id, [FromBody] PlanetChatChannel channel,
+        ILogger<PlanetChatChannel> logger)
+    {
+        // Get resources
+        var db = ctx.GetDb();
+        var member = ctx.GetMember();
+
+        if (channel.Planet_Id != planet_id)
+            return Results.BadRequest("Planet_Id mismatch.");
+
+        var nameValid = ValidateName(channel.Name);
+        if (!nameValid.Success)
+            return Results.BadRequest(nameValid.Message);
+
+        var descValid = ValidateDescription(channel.Description);
+        if (!descValid.Success)
+            return Results.BadRequest(descValid.Message);
+
+        var positionValid = await ValidateParentAndPosition(db, channel);
+        if (!positionValid.Success)
+            return Results.BadRequest(positionValid.Message);
+
+        // Ensure user has permission for parent category management
+        if (channel.Parent_Id is not null)
+        {
+            var parent_cat = await db.PlanetCategoryChannels.FindAsync(channel.Parent_Id);
+            if (!await parent_cat.HasPermissionAsync(member, CategoryPermissions.ManageCategory, db))
+                return ValourResult.LacksPermission(CategoryPermissions.ManageCategory);
+        }
+
+        try
+        {
+            await db.PlanetChatChannels.AddAsync(channel);
+            await db.SaveChangesAsync();
+        }
+        catch(System.Exception e)
+        {
+            logger.LogError(e.Message);
+            return Results.Problem(e.Message);
+        }
+
+        PlanetHub.NotifyPlanetItemChange(channel);
+
+        return Results.Created(channel.GetUri(), channel);
+    }
+
+    [ValourRoute(HttpVerbs.Put), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    [PlanetPermsRequired(PlanetPermissionsEnum.ManageChannels),
+     ChatChannelPermsRequired(ChatChannelPermissionsEnum.View,
+                              ChatChannelPermissionsEnum.ManageChannel)]
+    public static async Task<IResult> PutRouteAsync(HttpContext ctx, ulong id, [FromBody] PlanetChatChannel channel,
+        ILogger<PlanetChatChannel> logger)
+    {
+        // Get resources
+        var db = ctx.GetDb();
+        var old = ctx.GetItem<PlanetChatChannel>(id);
+
+        // Validation
+        if (old.Id != channel.Id)
+            return Results.BadRequest("Cannot change Id.");
+        if (old.Planet_Id != channel.Planet_Id)
+            return Results.BadRequest("Cannot change Planet_Id.");
+
+        var nameValid = ValidateName(channel.Name);
+        if (!nameValid.Success)
+            return Results.BadRequest(nameValid.Message);
+
+        var descValid = ValidateDescription(channel.Description);
+        if (!descValid.Success)
+            return Results.BadRequest(descValid.Message);
+
+        var positionValid = await ValidateParentAndPosition(db, channel);
+        if (!positionValid.Success)
+            return Results.BadRequest(positionValid.Message);
+
+        // Update
+        try
+        {
+            db.PlanetChatChannels.Update(channel);
+            await db.SaveChangesAsync();
+        }
+        catch (System.Exception e)
+        {
+            logger.LogError(e.Message);
+            return Results.Problem(e.Message);
+        }
+
+        PlanetHub.NotifyPlanetItemChange(channel);
+
+        // Response
+        return Results.Ok(channel);
+    }
+
+    [ValourRoute(HttpVerbs.Delete), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    [PlanetPermsRequired(PlanetPermissionsEnum.ManageChannels),
+     ChatChannelPermsRequired(ChatChannelPermissionsEnum.View,
+                              ChatChannelPermissionsEnum.ManageChannel)]
+    public static async Task<IResult> DeleteRouteAsync(HttpContext ctx, ulong id, ulong planet_id,
+        ILogger<PlanetChatChannel> logger)
+    {
+        var db = ctx.GetDb();
+        var channel = ctx.GetItem<PlanetChatChannel>(id);
+
+        // Always use transaction for multi-step DB operations
+        using var transaction = await db.Database.BeginTransactionAsync();
+
+        try
+        {
+            await channel.DeleteAsync(db);
+            await db.SaveChangesAsync();
+            await transaction.CommitAsync();
+        }
+        catch (System.Exception e)
+        {
+            logger.LogError(e.Message);
+            await transaction.RollbackAsync();
+            return Results.Problem(e.Message);
+        }
+
+        PlanetHub.NotifyPlanetItemDelete(channel);
+
+        return Results.NoContent();
+    }
+
+    // Message routes
+
+    [ValourRoute(HttpVerbs.Get, "/{id}/messages"), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    [ChatChannelPermsRequired(ChatChannelPermissionsEnum.ViewMessages)]
+    public static async Task<IResult> GetMessagesRouteAsync(HttpContext ctx, ulong id, ulong index = ulong.MaxValue, int count = 10)
+    {
+        if (count > 64)
+            return Results.BadRequest("Maximum count is 64.");
+
+        var channel = ctx.GetItem<PlanetChatChannel>(id);
+        var db = ctx.GetDb();
+
+        List<PlanetMessage> staged = PlanetMessageWorker.GetStagedMessages(id, count);
+
+        count = count - staged.Count;
+
+        if (count > 0)
+        {
+            var messages = await db.PlanetMessages.Where(x => x.Channel_Id == id && x.MessageIndex < index)
+                                                  .OrderByDescending(x => x.MessageIndex)
+                                                  .Take(count)
+                                                  .Reverse()
+                                                  .ToListAsync();
+
+            messages.AddRange(staged);
+
+            return Results.Json(messages);
+        }
+        else
+        {
+            return Results.Json(staged);
+        }
+    }
+
+    [ValourRoute(HttpVerbs.Post, "/{id}/messages"), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    [ChatChannelPermsRequired(ChatChannelPermissionsEnum.ViewMessages,
+                              ChatChannelPermissionsEnum.PostMessages)]
+    public static async Task<IResult> PostMessagesRouteAsync(HttpContext ctx, [FromBody] PlanetMessage message)
+    {
+        var member = ctx.GetMember();
+
+        if (message is null)
+            return Results.BadRequest("Include message in body.");
+
+        if (string.IsNullOrEmpty(message.Content) && string.IsNullOrEmpty(message.EmbedData))
+            return Results.BadRequest("Message content cannot be null");
+
+        if (message.Fingerprint is null)
+            return Results.BadRequest("Please include a Fingerprint.");
+
+        if (message.Author_Id != member.User_Id)
+            return Results.BadRequest("User_Id must match sender.");
+
+        if (message.Member_Id != member.Id)
+            return Results.BadRequest("Member_Id must match sender.");
+
+        if (message.Content != null && message.Content.Length > 2048)
+            return Results.BadRequest("Content must be under 2048 chars");
+
+
+        if (message.EmbedData != null && message.EmbedData.Length > 65535)
+            return Results.BadRequest("EmbedData must be under 65535 chars");
+
+        // Handle URL content
+        message.Content = await MPSUtils.HandleUrls(message.Content);
+        message.Id = IdManager.Generate();
+
+        PlanetMessageWorker.AddToQueue(message);
+
+        StatWorker.IncreaseMessageCount();
+
+        return Results.Ok();
+    }
+
+    [ValourRoute(HttpVerbs.Get, "/{id}/messages/{message_id}"), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    [ChatChannelPermsRequired(ChatChannelPermissionsEnum.ViewMessages)]
+    public static async Task<IResult> DeleteMessagesRouteAsync(HttpContext ctx, ulong id, ulong message_id,
+        ILogger<PlanetChatChannel> logger)
+    {
+        var db = ctx.GetDb();
+        var member = ctx.GetMember();
+        var channel = ctx.GetItem<PlanetChatChannel>(id);
+
+        var message = await FindAsync<PlanetMessage>(message_id, db);
+
+        var inDb = true;
+
+        if (message is null)
+        {
+            inDb = false;
+
+            // Try to find in staged
+            message = PlanetMessageWorker.GetStagedMessage(message_id);
+            if (message is null)
+                return ValourResult.NotFound<PlanetMessage>();
+        }
+
+        if (message.Channel_Id != id)
+            return ValourResult.NotFound<PlanetMessage>();
+
+        if (member.Id != message.Member_Id)
+        {
+            if (!await channel.HasPermissionAsync(member, ChatChannelPermissions.ManageMessages, db))
+                return ValourResult.LacksPermission(ChatChannelPermissions.ManageMessages);
+        }
+
+        // Remove from staging
+        PlanetMessageWorker.RemoveFromQueue(message);
+
+        // If in db, remove from db
+        if (inDb)
+        {
+            try
+            {
+                db.PlanetMessages.Remove(message);
+                await db.SaveChangesAsync();
+            }
+            catch (System.Exception e)
+            {
+                logger.LogError(e.Message);
+                return Results.Problem(e.Message);
+            }
+        }
+
+        PlanetHub.NotifyMessageDeletion(message);
+
+        return Results.NoContent();
+    }
+
+    #endregion
+
+    #region Validation
+
+    /// <summary>
+    /// Validates that a given name is allowable
     /// </summary>
     public static TaskResult ValidateName(string name)
     {
@@ -117,132 +433,33 @@ public class PlanetChatChannel : PlanetChannel, ISharedPlanetChatChannel, INodeS
     }
 
     /// <summary>
-    /// Returns all members who can see this channel
+    /// Validates that a given description is allowable
     /// </summary>
-    public async Task<List<PlanetMember>> GetChannelMembersAsync(ValourDB db = null)
+    public static TaskResult ValidateDescription(string desc)
     {
-        List<PlanetMember> members = new List<PlanetMember>();
-
-        bool createdb = false;
-        if (db == null) { db = new ValourDB(ValourDB.DBOptions); createdb = true; }
-
-        var planetMembers = db.PlanetMembers.Include(x => x.RoleMembership).Where(x => x.Planet_Id == Planet_Id);
-
-        foreach (var member in planetMembers)
+        if (desc.Length > 500)
         {
-            if (await HasPermission(member, ChatChannelPermissions.View, db))
-            {
-                members.Add(member);
-            }
+            return new TaskResult(false, "Planet descriptions must be 500 characters or less.");
         }
 
-        if (createdb) { await db.DisposeAsync(); }
-
-        return members;
+        return TaskResult.SuccessResult;
     }
 
-    #region API Methods
-
-    public override async Task<TaskResult> CanGetAsync(PlanetMember member, ValourDB db)
+    public static async Task<TaskResult> ValidateParentAndPosition(ValourDB db, PlanetChatChannel channel)
     {
-        if (member is null)
-            return new TaskResult(false, "User is not a member of the target planet");
-
-        if (!await HasPermission(member, ChatChannelPermissions.View, db))
-            return new TaskResult(false, "Member lacks channel permission " + ChatChannelPermissions.View.Name);
-
-        return new TaskResult(true, "Success");
-    }
-
-    public override async Task<TaskResult> CanDeleteAsync(PlanetMember member, ValourDB db)
-    {
-        // Needs to be able to GET in order to do anything else
-        var canGet = await CanGetAsync(member, db);
-        if (!canGet.Success)
-            return canGet;
-
-        await GetPlanetAsync(db);
-
-        if (!await Planet.HasPermissionAsync(member, PlanetPermissions.ManageChannels, db))
-            return new TaskResult(false, "Member lacks planet permission " + PlanetPermissions.ManageChannels.Name);
-
-        if (!await HasPermission(member, ChatChannelPermissions.ManageChannel, db))
-            return new TaskResult(false, "Member lacks channel permission " + ChatChannelPermissions.ManageChannel.Name);
-
-
-        return new TaskResult(true, "Success");
-    }
-
-    public override async Task<TaskResult> CanUpdateAsync(PlanetMember member, PlanetItem old, ValourDB db)
-    {
-        // Similar to Create but also needs specific channel perms
-        var canCreate = await CanCreateAsync(member, db);
-        if (!canCreate.Success)
-            return canCreate;
-
-        if (!await HasPermission(member, ChatChannelPermissions.ManageChannel, db))
-            return new TaskResult(false, "Member lacks channel permission " + ChatChannelPermissions.ManageChannel.Name);
-
-        return new TaskResult(true, "Success");
-    }
-
-    public override async Task<TaskResult> CanCreateAsync(PlanetMember member, ValourDB db)
-    {
-        await GetPlanetAsync(db);
-
-        if (!await Planet.HasPermissionAsync(member, PlanetPermissions.ManageChannels, db))
-            return new TaskResult(false, "Member lacks planet permission " + PlanetPermissions.ManageChannels.Name);
-
-        var valid = await ValidateAsync(db);
-        if (!valid.Success)
-            return valid;
-
-        return new TaskResult(true, "Success");
-    }
-
-    public override async Task DeleteAsync(ValourDB db)
-    {
-        // Remove permission nodes
-        db.PermissionsNodes.RemoveRange(
-            db.PermissionsNodes.Where(x => x.Target_Id == Id)
-        );
-
-        // Remove messages
-        db.PlanetMessages.RemoveRange(
-            db.PlanetMessages.Where(x => x.Channel_Id == Id)
-        );
-
-        // Remove channel
-        db.PlanetChatChannels.Remove(
-            await db.PlanetChatChannels.FirstOrDefaultAsync(x => x.Id == Id)
-        );
-
-        // Save changes
-        await db.SaveChangesAsync();
-
-        // Notify channel deletion
-        PlanetHub.NotifyPlanetItemChange(this);
-    }
-
-    public async Task<TaskResult> ValidateAsync(ValourDB db)
-    {
-        var nameValid = ValidateName(Name);
-        if (!nameValid.Success)
-            return nameValid;
-
-        if (Description.Length > 128)
-            return new TaskResult(false, "Description must be at or under 128 characters");
-
         // Logic to check if parent is legitimate
-        if (Parent_Id is not null)
+        if (channel.Parent_Id is not null)
         {
             var parent = await db.PlanetCategoryChannels.FirstOrDefaultAsync
-                (x => x.Id == Parent_Id
-                && x.Planet_Id == Planet_Id); // This ensures the result has the same planet id
+                (x => x.Id == channel.Parent_Id
+                && x.Planet_Id == channel.Planet_Id); // This ensures the result has the same planet id
 
             if (parent is null)
                 return new TaskResult(false, "Parent ID is not valid");
         }
+
+        if (!await HasUniquePosition(db, channel))
+            return new TaskResult(false, "The position is already taken.");
 
         return new TaskResult(true, "Valid");
     }
