@@ -1,9 +1,17 @@
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Text.Json.Serialization;
+using System.Web.Mvc;
+using Valour.Database.Attributes;
+using Valour.Database.Extensions;
+using Valour.Database.Items.Authorization;
 using Valour.Shared;
 using Valour.Shared.Authorization;
+using Valour.Shared.Http;
 using Valour.Shared.Items;
 using Valour.Shared.Items.Planets.Members;
 
@@ -18,7 +26,7 @@ namespace Valour.Database.Items.Planets.Members;
 /// <summary>
 /// This represents a user within a planet and is used to represent membership
 /// </summary>
-public class PlanetBan : PlanetItem, ISharedPlanetBan, INodeSpecific
+public class PlanetBan : PlanetItem, ISharedPlanetBan
 {
     /// <summary>
     /// The member that banned the user
@@ -55,92 +63,175 @@ public class PlanetBan : PlanetItem, ISharedPlanetBan, INodeSpecific
     /// </summary>
     public bool Permanent => Expires == null;
 
-    /// <summary>
-    /// Creates this item in the database and bans the user
-    /// </summary>
-    public override async Task CreateAsync(ValourDB db)
+    #region Routes
+
+    [ValourRoute(HttpVerbs.Get), TokenRequired, InjectDB]
+    [PlanetMembershipRequired]
+    //[PlanetPermsRequired(PlanetPermissionsEnum.Ban)] (There is an exception to this!)
+    public static async Task<IResult> GetRoute(HttpContext ctx, ulong id)
     {
-        PlanetMember target = await db.PlanetMembers.FirstOrDefaultAsync(x => x.Id == Target_Id);
+        var db = ctx.GetDb();
+        var ban = await FindAsync<PlanetBan>(id, db);
+        var member = ctx.GetMember();
 
-        var roles = db.PlanetRoleMembers.Where(x => x.Member_Id == target.Id && x.Planet_Id == Planet_Id);
+        if (ban is null)
+            return ValourResult.NotFound<PlanetBan>();
 
-        await db.PlanetRoleMembers.BulkDeleteAsync(roles);
-
-        db.PlanetMembers.Remove(target);
-
-        await db.AddAsync(this);
-        await db.SaveChangesAsync();
-
-        foreach (var role in roles)
+        // You can retrieve your own ban
+        if (ban.Target_Id != member.Id)
         {
-            PlanetHub.NotifyPlanetItemDelete(role);
+            if (!await member.HasPermissionAsync(PlanetPermissions.Ban, db))
+                return ValourResult.LacksPermission(PlanetPermissions.Ban);
         }
 
-        PlanetHub.NotifyPlanetItemDelete(target);
-        PlanetHub.NotifyPlanetItemChange(this);
+        return Results.Json(ban);
     }
 
-    public override async Task<TaskResult> CanGetAsync(PlanetMember member, ValourDB db)
+    [ValourRoute(HttpVerbs.Post), TokenRequired, InjectDB]
+    [PlanetMembershipRequired, PlanetPermsRequired(PlanetPermissionsEnum.Ban)]
+    public static async Task<IResult> PostRoute(HttpContext ctx, ulong planet_id, [FromBody] PlanetBan ban,
+        ILogger<PlanetBan> logger)
     {
-        // Members can get their own ban info
-        if (member.Id == Target_Id)
-            return TaskResult.SuccessResult;
+        var db = ctx.GetDb();
+        var member = ctx.GetMember();
 
-        return await CanBanAsync(member, db);
-    }
+        if (ban is null)
+            return Results.BadRequest("Include ban in body.");
 
-    public override async Task<TaskResult> CanDeleteAsync(PlanetMember member, ValourDB db) 
-        => await CanBanAsync(member, db);
+        if (ban.Planet_Id != planet_id)
+            return Results.BadRequest("Planet_Id mismatch.");
 
-    public override async Task<TaskResult> CanUpdateAsync(PlanetMember member, PlanetItem old, ValourDB db)
-    {
-        TaskResult canBan = await CanBanAsync(member, db);
-        if (!canBan.Success)
-            return canBan;
+        if (ban.Banner_Id != member.Id)
+            return Results.BadRequest("Banner_Id should match member Id.");
 
-        var oldBan = old as PlanetBan;
+        if (ban.Target_Id == member.Id)
+            return Results.BadRequest("You cannot ban yourself.");
 
-        if (this.Target_Id != oldBan.Target_Id)
-            return new TaskResult(false, "You cannot change who was banned");
+        // Ensure it doesn't already exist
+        if (await db.PlanetBans.AnyAsync(x => x.Planet_Id == ban.Planet_Id && x.Target_Id == ban.Target_Id))
+            return Results.BadRequest("Ban already exists for user.");
 
-        if (this.Banner_Id != oldBan.Banner_Id)
-            return new TaskResult(false, "You cannot change who banned the user");
-
-        if (this.Time != oldBan.Time)
-            return new TaskResult(false, "You cannot change the creation time");
-
-        return TaskResult.SuccessResult;
-    }
-
-    public override async Task<TaskResult> CanCreateAsync(PlanetMember member, ValourDB db)
-    {
-        TaskResult canBan = await CanBanAsync(member, db);
-        if (!canBan.Success)
-            return canBan;
-
-        // Ensure target exists
-        PlanetMember target = await db.PlanetMembers.FirstOrDefaultAsync(x => x.User_Id == Target_Id);
+        // Ensure user has more authority than the user being banned
+        var target = await PlanetMember.FindAsync(ban.Target_Id, planet_id, db);
 
         if (target is null)
-            return new TaskResult(false, $"Target not found");
+            return ValourResult.NotFound<PlanetMember>();
 
-        if (Banner_Id == member.Id)
-            return new TaskResult(false, $"You cannot ban yourself");
+        if (await target.GetAuthorityAsync(db) >= await member.GetAuthorityAsync(db))
+            return ValourResult.Forbid("The target has a higher authority than you.");
 
-        if (await target.GetAuthorityAsync() >= await member.GetAuthorityAsync())
-            return new TaskResult(false, "You cannot ban users with higher or same authority than your own.");
+        var tran = await db.Database.BeginTransactionAsync();
 
-        if (Banner_Id != member.Id)
-            return new TaskResult(false, $"The banner is not the same as the auth member");
+        try
+        {
+            // Add ban
+            await db.PlanetBans.AddAsync(ban);
 
-        // Set time from server
-        Time = DateTime.UtcNow;
+            // Save changes
+            await db.SaveChangesAsync();
 
-        return TaskResult.SuccessResult;
+            // Delete target member
+            await target.DeleteAsync(db);
+        }
+        catch(System.Exception e)
+        {
+            logger.LogError(e.Message);
+            await tran.RollbackAsync();
+            return Results.Problem(e.Message);
+        }
+
+        await tran.CommitAsync();
+
+        // Notify of changes
+        PlanetHub.NotifyPlanetItemChange(ban);
+        PlanetHub.NotifyPlanetItemDelete(target);
+
+        return Results.Created(ban.GetUri(), ban);
     }
-    
-    private static async Task<TaskResult> CanBanAsync(PlanetMember member, ValourDB db) 
-        => !await member.HasPermissionAsync(PlanetPermissions.Ban, db)
-            ? new TaskResult(false, "Member lacks Planet Permission " + PlanetPermissions.Ban.Name)
-            : TaskResult.SuccessResult;
+
+    [ValourRoute(HttpVerbs.Put), TokenRequired, InjectDB]
+    [PlanetMembershipRequired, PlanetPermsRequired(PlanetPermissionsEnum.Ban)]
+    public static async Task<IResult> PutRoute(HttpContext ctx, ulong id, ulong planet_id, [FromBody] PlanetBan ban,
+        ILogger<PlanetBan> logger)
+    {
+        var db = ctx.GetDb();
+        var member = ctx.GetMember();
+
+        if (ban is null)
+            return Results.BadRequest("Include updated ban in body.");
+
+        var old = await FindAsync<PlanetBan>(id, db);
+
+        if (old is null)
+            return ValourResult.NotFound<PlanetBan>();
+
+        if (ban.Planet_Id != old.Planet_Id)
+            return Results.BadRequest("You cannot change the Planet_Id.");
+
+        if (ban.Target_Id != old.Target_Id)
+            return Results.BadRequest("You cannot change who was banned.");
+
+        if (ban.Banner_Id != old.Banner_Id)
+            return Results.BadRequest("You cannot change who banned the user.");
+
+        if (ban.Time != old.Time)
+            return Results.BadRequest("You cannot change the creation time");
+
+        try
+        {
+            db.PlanetBans.Update(ban);
+            await db.SaveChangesAsync();
+        }
+        catch (System.Exception e)
+        {
+            logger.LogError(e.Message);
+            return Results.Problem(e.Message);
+        }
+
+        // Notify of changes
+        PlanetHub.NotifyPlanetItemChange(ban);
+
+        return Results.Ok(ban);
+    }
+
+    [ValourRoute(HttpVerbs.Delete), TokenRequired, InjectDB]
+    [PlanetMembershipRequired, PlanetPermsRequired(PlanetPermissionsEnum.Ban)]
+    public static async Task<IResult> DeleteRoute(HttpContext ctx, ulong id, ulong planet_id,
+        ILogger<PlanetBan> logger)
+    {
+        var db = ctx.GetDb();
+        var member = ctx.GetMember();
+
+        var ban = await FindAsync<PlanetBan>(id, db);
+
+        // Ensure the user unbanning is either the user that made the ban, or someone
+        // with equal or higher authority to them
+
+        if (ban.Banner_Id != member.Id)
+        {
+            var banner = await FindAsync<PlanetMember>(ban.Banner_Id, db);
+
+            if (await banner.GetAuthorityAsync(db) > await member.GetAuthorityAsync(db))
+                return ValourResult.Forbid("The banner of this user has higher authority than you.");
+        }
+
+        try
+        {
+            db.PlanetBans.Remove(ban);
+            await db.SaveChangesAsync();
+        }
+        catch (System.Exception e)
+        {
+            logger.LogError(e.Message);
+            return Results.Problem(e.Message);
+        }
+        
+
+        // Notify of changes
+        PlanetHub.NotifyPlanetItemDelete(ban);
+
+        return Results.NoContent();
+    }
+
+    #endregion
 }
