@@ -1,5 +1,10 @@
-﻿using System.ComponentModel.DataAnnotations.Schema;
+﻿using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Text.Json.Serialization;
+using Valour.Database.Items.Authorization;
 using Valour.Database.Items.Users;
 using Valour.Shared;
 using Valour.Shared.Authorization;
@@ -14,7 +19,7 @@ using Valour.Shared.Items.Planets.Members;
 
 namespace Valour.Database.Items.Planets.Members;
 
-public class PlanetRoleMember : PlanetItem, ISharedPlanetRoleMember, INodeSpecific
+public class PlanetRoleMember : PlanetItem, ISharedPlanetRoleMember
 {
     [ForeignKey("Member_Id")]
     [JsonIgnore]
@@ -34,52 +39,7 @@ public class PlanetRoleMember : PlanetItem, ISharedPlanetRoleMember, INodeSpecif
 
     public override ItemType ItemType => ItemType.PlanetRoleMember;
 
-    public override async Task<TaskResult> CanCreateAsync(PlanetMember member, ValourDB db)
-    {
-        // Needs to be able to GET in order to do anything else
-        var canGet = await CanGetAsync(member, db);
-        if (!canGet.Success)
-            return canGet;
-
-        if (!await member.HasPermissionAsync(PlanetPermissions.ManageRoles, db))
-            return new TaskResult(false, "You lack the Planet Permission " + PlanetPermissions.ManageRoles.Name);
-
-        var role = await db.PlanetRoles.FindAsync(Role_Id);
-
-        if (role is null || role.Planet_Id != Planet_Id)
-            return new TaskResult(false, "The Role_Id is invalid.");
-
-        var memberAuthority = await member.GetAuthorityAsync();
-
-        if (role.GetAuthority() >= memberAuthority)
-            return new TaskResult(false, "You have less authority than the role you are trying to modify.");
-
-        var targetMember = await db.PlanetMembers.FindAsync(Member_Id);
-
-        if (targetMember is null)
-            return new TaskResult(false, "Member not found.");
-
-        if (targetMember.Planet_Id != Planet_Id)
-            return new TaskResult(false, "Member Planet_Id mismatch.");
-
-        if (targetMember.User_Id != User_Id)
-            return new TaskResult(false, "Member User_Id mismatch.");
-
-        if (role.Planet_Id != Planet_Id)
-            return new TaskResult(false, "Role Planet_Id mismatch.");
-
-        return TaskResult.SuccessResult;
-    }
-
-    public override async Task<TaskResult> CanDeleteAsync(PlanetMember member, ValourDB db)
-    {
-        // Same permissions
-        return await CanCreateAsync(member, db);
-    } 
-
-    public override async Task<TaskResult> CanUpdateAsync(PlanetMember member, PlanetItem old, ValourDB db)
-    {
-        return await Task.FromResult(new TaskResult(false, "You cannot modify this object.")); 
-    }
+    // This doesn't really even need API routes, it's just used internally to map roles to members.
+    // Use a route from PlanetMember if you need to get someone's roles.
 }
 
