@@ -13,6 +13,7 @@ using Valour.Database.Items.Authorization;
 using Valour.Shared.Authorization;
 using Valour.Server.Users;
 using Valour.Database.Items.Planets.Members;
+using Valour.Shared.Http;
 
 namespace Valour.Server.API
 {
@@ -41,18 +42,18 @@ namespace Valour.Server.API
             app.MapPost("api/user/verify/{code}", VerifyEmail);
         }
 
-        private static async Task GetUser(HttpContext ctx, ValourDB db, ulong user_id, [FromHeader] string authorization)
+        private static async Task<IResult> GetUser(HttpContext ctx, ValourDB db, ulong user_id, [FromHeader] string authorization)
         {
             var authToken = await AuthToken.TryAuthorize(authorization, db);
-
-            if (authToken == null) { await TokenInvalid(ctx); return; }
+            if (authToken == null)
+                return ValourResult.NoToken();
 
             var user = await db.Users.FindAsync(user_id);
 
-            if (user == null) { await NotFound("User not found", ctx); return; }
+            if (user == null)
+                return ValourResult.NotFound<User>();
 
-            ctx.Response.StatusCode = 200;
-            await ctx.Response.WriteAsJsonAsync(user);
+            return Results.Json(user);
         }
 
         private static async Task VerifyEmail(HttpContext ctx, ValourDB db, string code)
