@@ -8,7 +8,8 @@ namespace Valour.Shared.Items;
 public interface ISharedItem
 {
     ulong Id { get; set; }
-    abstract ItemType ItemType { get; }
+    ItemType ItemType { get; }
+    string Node { get; }
 }
 
 
