@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Valour.Server.MPS
+namespace Valour.Shared.MPS
 {
     public class MPSConfig
     {
